💄 UI: 공개/비공개 토글 버튼 공용 컴포넌트화
