+++ src/pages/CreatePinPage/CreatePinPage.jsx
@@ -1,6 +1,5 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
-import ReactSwitch from 'react-switch';
 import arrowDown from '../../assets/images/MusicSearchPage/arrow_down.svg';
 import SideSection from '../../components/common/SideSection';
 import SearchBar from '../../components/MusicSearchPage/SearchPage/SearchBar';
@@ -10,17 +9,13 @@ import PinComponent from '../../components/MusicSearchPage/PinComponent';
 import { GenreList } from '../../constants/GenreList';
 import { ReactComponent as Calendar} from '../../assets/images/CreatePin/calendar_month.svg';
 import { ReactComponent as Location} from '../../assets/images/CreatePin/location_on.svg';
+import PublicToggle from '../../components/common/PublicToggle';
 
 const CreatePinPage = () => {
     let [inputCount, setInputCount] = useState(0);
-    const [isPublic, setIsPublic] = useState(false);
 
     const onInputHandler = (e) => {
         setInputCount(e.target.value.length);
-    };
-
-    const handleToggle = (checked) => {
-        setIsPublic(checked);
     };
 
     return (
@@ -49,25 +44,7 @@ const CreatePinPage = () => {
                 </TextNum>
                 <IsPublic>
                     <Title>공개 여부</Title>
-                    <Switch>
-                        <ToggleText>{isPublic ? "공개" : "비공개"}</ToggleText>
-                        <StyledSwitch
-                            onChange={handleToggle}
-                            checked={isPublic}
-                            onColor='#FFFFFF'
-                            onHandleColor='#000000'
-                            offColor='#000000'
-                            offHandleColor='#FFFFFF'
-                            handleDiameter={20}
-                            uncheckedIcon={false}
-                            checkedIcon={false}
-                            // boxShadow="0px 1px 5px rgba(0, 0, 0, 0.6)"
-                            // activeBoxShadow="0px 0px 1px 10px rgba(0, 0, 0, 0.2)"
-                            height={30}
-                            width={60}
-                            className="react-switch"
-                        />
-                    </Switch>
+                    <PublicToggle />
                 </IsPublic>
                 <CreateBtn>핀 생성하기</CreateBtn>
         </SideSection>
@@ -197,28 +174,6 @@ const IsPublic = styled.div`
     display: flex;
     justify-content: space-between;
     align-items: center;
-`;
-
-const Switch = styled.div`
-    display: flex;
-    align-items: center;
-    margin-right: 20px;
-`;
-
-const ToggleText = styled.span`
-    margin-right: 10px;
-    font-family: Pretendard;
-    font-size: 16px;
-    font-style: normal;
-    font-weight: 400;
-    color: var(--gray02, #747474);
-    margin-right: 16px;
-`;
-
-const StyledSwitch = styled(ReactSwitch)`
-    &.react-switch {
-        border: ${({ checked }) => (checked ? '1px solid #000000' : 'none')};
-    }
 `;
 
 const CreateBtn = styled.button`
